pyproject.toml
[tool.poetry]
name = "stock-database"
version = "0.1.0"
description = ""
authors = ["Tan Yong Sheng"]
license = "MIT license"
readme = "README.md"
packages = [{include = "stock_database"}]

[tool.poetry.dependencies]
python = "^3.8"
sqlmodel = "^0.0.8"
python-dateutil = "^2.8.2"
newsapi-python = "^0.2.7"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.0"
pylint-pytest = "^1.1.2"
chainlit = "^0.6.0"
pandas-stubs = "^2.0.2.230605"
types-python-dateutil = "^2.8.19.13"
financedatabase = "^2.1.0"
aiohttp = "^3.8.5"
aiosignal = "^1.3.1"
anyio = "^3.7.1"
astroid = "^2.15.6"
async-timeout = "^4.0.2"
asyncer = "^0.0.2"
attrs = "^23.1.0"
auth0-python = "^4.3.0"
backoff = "^2.2.1"
bidict = "^0.22.1"
certifi = "^2023.5.7"
cffi = "^1.15.1"
charset-normalizer = "^3.2.0"
click = "^8.1.6"
colorama = "^0.4.6"
cryptography = "^41.0.2"
deprecated = "^1.2.14"
dill = "^0.3.6"
exceptiongroup = "^1.1.2"
fastapi-socketio = "^0.0.10"
filetype = "^1.2.0"
frozenlist = "^1.4.0"
googleapis-common-protos = "^1.59.1"
greenlet = "^2.0.2"
grpcio = "^1.56.2"
h11 = "^0.14.0"
httpcore = "^0.17.3"
httpx = "^0.24.1"
idna = "^3.4"
iniconfig = "^2.0.0"
isort = "^5.12.0"
jinja2 = "^3.1.2"
lazy-object-proxy = "^1.9.0"
markupsafe = "^2.1.3"
marshmallow-enum = "^1.5.1"
marshmallow = "^3.20.1"
mccabe = "^0.7.0"
multidict = "^6.0.4"
mypy-extensions = "^1.0.0"
nest-asyncio = "^1.5.6"
langchain = "^0.0.239"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
mypy = "^1.4.1"
black = "^23.7.0"
ruff = "^0.0.278"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
